+++ azure-iot-device/azure/iot/device/common/transport/mqtt/pipeline_events_mqtt.py
@@ -3,7 +3,7 @@
 # Licensed under the MIT License. See License.txt in the project root for
 # license information.
 # --------------------------------------------------------------------------
-from .pipeline_events_base import PipelineEvent
+from azure.iot.device.common.transport.pipeline_events_base import PipelineEvent
 
 
 class IncomingMessage(PipelineEvent):
