+++ azure-iot-device/azure/iot/device/common/transport/mqtt/pipeline_stages_mqtt.py
@@ -5,10 +5,10 @@
 # --------------------------------------------------------------------------
 
 import logging
-from . import pipeline_ops_base
+from azure.iot.device.common.transport.pipeline_stages_base import PipelineStage
+from azure.iot.device.common.transport import pipeline_ops_base
 from . import pipeline_ops_mqtt
 from . import pipeline_events_mqtt
-from .pipeline_stages_base import PipelineStage
 from azure.iot.device.common.transport.mqtt.mqtt_provider import MQTTProvider
 
 logger = logging.getLogger(__name__)
