+++ azure-iot-device/azure/iot/device/common/transport/pipeline_stages_base.py
