+++ azure-iot-device/azure/iot/device/iothub/transport/constant.py
