+++ azure-iot-device/azure/iot/device/iothub/transport/mqtt/__init__.py
@@ -1,0 +1,8 @@
+"""Azure IoT Hub Device SDK MQTT Transport
+
+This package provides the MQTT transport for use with the Azure IoT Hub Device SDK.
+
+INTERNAL USAGE ONLY
+"""
+
+from .mqtt_transport import MQTTTransport
