+++ azure-iot-device/azure/iot/device/iothub/transport/mqtt/pipeline_events_iothub.py
@@ -2,7 +2,7 @@
 # Licensed under the MIT License. See License.txt in the project root for
 # license information.
 # --------------------------------------------------------------------------
-from .pipeline_events_base import PipelineEvent
+from azure.iot.device.common.transport.pipeline_events_base import PipelineEvent
 
 
 class C2DMessage(PipelineEvent):
