+++ azure-iot-device/tests/iothub/aio/test_async_clients.py
@@ -9,10 +9,10 @@ import asyncio
 import abc
 import six
 from azure.iot.device.iothub.aio import IoTHubDeviceClient, IoTHubModuleClient
-from azure.iot.device.common.transport.mqtt import MQTTTransport
+from azure.iot.device.iothub.transport.mqtt import MQTTTransport
 from azure.iot.device.iothub import Message
 from azure.iot.device.iothub.aio.async_inbox import AsyncClientInbox
-from azure.iot.device.common.transport import constant
+from azure.iot.device.iothub.transport import constant
 
 # auth_provider and transport fixtures are implicitly included
